# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hacknation Reasoning Auditor API",
    "description": "FastAPI service for Person B's v2 reasoning layer.",
    "version": "0.0.2"
  },
  "paths": {
    "/api/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Query",
        "operationId": "query_api_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/facility/{facility_id}": {
      "get": {
        "tags": [
          "facility"
        ],
        "summary": "Facility",
        "operationId": "facility_api_facility__facility_id__get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityTrustRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/facility/{facility_id}/evidence": {
      "get": {
        "tags": [
          "facility"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_facility__facility_id__evidence_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityEvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/desert": {
      "get": {
        "tags": [
          "desert"
        ],
        "summary": "Desert",
        "operationId": "desert_api_desert_get",
        "parameters": [
          {
            "name": "capability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "pin_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pin Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PinCodeDesert"
                  },
                  "title": "Response Desert Api Desert Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/desert/summary": {
      "get": {
        "tags": [
          "desert"
        ],
        "summary": "Summary",
        "operationId": "summary_api_desert_summary_get",
        "parameters": [
          {
            "name": "capability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesertSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trace/{trace_id}/timeline": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Timeline",
        "operationId": "timeline_api_trace__trace_id__timeline_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilityClaim": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "claim_present": {
            "type": "boolean",
            "title": "Claim Present"
          },
          "self_consistency_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Self Consistency Score"
          },
          "coherence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Coherence Score"
          },
          "peer_anomaly_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Peer Anomaly Score"
          },
          "trust_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Trust Score"
          },
          "confidence_interval_low": {
            "type": "number",
            "title": "Confidence Interval Low"
          },
          "confidence_interval_high": {
            "type": "number",
            "title": "Confidence Interval High"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "claim_present",
          "self_consistency_score",
          "coherence_score",
          "peer_anomaly_score",
          "trust_score",
          "confidence_interval_low",
          "confidence_interval_high",
          "citations"
        ],
        "title": "CapabilityClaim",
        "description": "One row per (facility, capability). The atomic unit of trust."
      },
...
      "Citation": {
        "properties": {
          "source_field": {
            "type": "string",
            "title": "Source Field"
          },
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "char_start": {
            "type": "integer",
            "title": "Char Start"
          },
          "char_end": {
            "type": "integer",
            "title": "Char End"
          }
        },
        "type": "object",
        "required": [
          "source_field",
          "sentence",
          "char_start",
          "char_end"
        ],
        "title": "Citation",
        "description": "Exact provenance for any extracted fact."
      },
...
      "DesertGroupSummary": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "critical_pin_count": {
            "type": "integer",
            "title": "Critical Pin Count"
          },
          "population_at_risk": {
            "type": "integer",
            "title": "Population At Risk"
          },
          "average_desert_score": {
            "type": "number",
            "title": "Average Desert Score"
          }
        },
        "type": "object",
        "required": [
          "key",
          "row_count",
          "critical_pin_count",
          "population_at_risk",
          "average_desert_score"
        ],
        "title": "DesertGroupSummary"
      },
      "DesertPinSummary": {
        "properties": {
          "pin_code": {
            "type": "string",
            "title": "Pin Code"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "population": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "desert_score": {
            "type": "number",
            "title": "Desert Score"
          },
          "nearest_verified_facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nearest Verified Facility Id"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          }
        },
        "type": "object",
        "required": [
          "pin_code",
          "state",
          "district",
          "capability",
          "desert_score"
        ],
        "title": "DesertPinSummary"
      },
      "DesertSummary": {
        "properties": {
          "capability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "critical_pin_count": {
            "type": "integer",
            "title": "Critical Pin Count"
          },
          "population_at_risk": {
            "type": "integer",
            "title": "Population At Risk"
          },
          "average_desert_score": {
            "type": "number",
            "title": "Average Desert Score"
          },
          "top_deserts": {
            "items": {
              "$ref": "#/components/schemas/DesertPinSummary"
            },
            "type": "array",
            "title": "Top Deserts"
          },
          "by_state": {
            "items": {
              "$ref": "#/components/schemas/DesertGroupSummary"
            },
            "type": "array",
            "title": "By State"
          },
          "by_capability": {
            "items": {
              "$ref": "#/components/schemas/DesertGroupSummary"
            },
            "type": "array",
            "title": "By Capability"
          }
        },
        "type": "object",
        "required": [
          "total_rows",
          "critical_pin_count",
          "population_at_risk",
          "average_desert_score",
          "top_deserts",
          "by_state",
          "by_capability"
        ],
        "title": "DesertSummary"
      },
...
      "EvidenceSnippet": {
        "properties": {
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "source_field": {
            "type": "string",
            "title": "Source Field"
          },
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "char_start": {
            "type": "integer",
            "title": "Char Start"
          },
          "char_end": {
            "type": "integer",
            "title": "Char End"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "facility_id",
          "capability",
          "source_field",
          "sentence",
          "char_start",
          "char_end"
        ],
        "title": "EvidenceSnippet"
      },
...
      "FacilityEvidenceResponse": {
        "properties": {
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "facility_name": {
            "type": "string",
            "title": "Facility Name"
          },
          "pin_code": {
            "type": "string",
            "title": "Pin Code"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "facility_type": {
            "type": "string",
            "title": "Facility Type"
          },
          "overall_trust_score": {
            "type": "number",
            "title": "Overall Trust Score"
          },
          "has_flags": {
            "type": "boolean",
            "title": "Has Flags"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          },
          "flag_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag Summary"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalBreakdown"
            },
            "type": "array",
            "title": "Signals"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceSnippet"
            },
            "type": "array",
            "title": "Evidence"
          },
          "audit_summary": {
            "type": "string",
            "title": "Audit Summary"
          }
        },
        "type": "object",
        "required": [
          "facility_id",
          "facility_name",
          "pin_code",
          "state",
          "district",
          "facility_type",
          "overall_trust_score",
          "has_flags",
          "flags",
          "signals",
          "evidence",
          "audit_summary"
        ],
        "title": "FacilityEvidenceResponse"
      },
...
      "FacilityTrustRecord": {
        "properties": {
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "facility_name": {
            "type": "string",
            "title": "Facility Name"
          },
          "pin_code": {
            "type": "string",
            "title": "Pin Code"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "facility_type": {
            "type": "string",
            "title": "Facility Type"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CapabilityClaim"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "overall_trust_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Trust Score"
          },
          "extraction_run_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extraction Run Ids"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "facility_id",
          "facility_name",
          "pin_code",
          "state",
          "district",
          "lat",
          "lon",
          "facility_type",
          "capabilities",
          "overall_trust_score",
          "extraction_run_ids",
          "last_updated"
        ],
        "title": "FacilityTrustRecord",
        "description": "One row per facility in `gold.facility_trust`."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PinCodeDesert": {
        "properties": {
          "pin_code": {
            "type": "string",
            "title": "Pin Code"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "population": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "nearest_verified_facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nearest Verified Facility Id"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "desert_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Desert Score"
          }
        },
        "type": "object",
        "required": [
          "pin_code",
          "state",
          "district",
          "lat",
          "lon",
          "capability",
          "desert_score"
        ],
        "title": "PinCodeDesert",
        "description": "One row per (pin_code, capability) in `gold.pin_code_desert`."
      },
...
      "QueryRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "user_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Lat"
          },
          "user_lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Lon"
          },
          "max_distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Distance Km"
          },
          "min_trust_score": {
            "type": "number",
            "title": "Min Trust Score",
            "default": 0.5
          },
          "capabilities_filter": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capabilities Filter"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "QueryRequest",
        "description": "User query into the Coordinator agent."
      },
      "QueryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/FacilityTrustRecord"
            },
            "type": "array",
            "title": "Candidates"
          },
          "citations_per_candidate": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Citation"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Citations Per Candidate"
          },
          "critic_verdict": {
            "type": "string",
            "enum": [
              "supported",
              "partial",
              "unsupported"
            ],
            "title": "Critic Verdict"
          },
          "critic_reasoning": {
            "type": "string",
            "title": "Critic Reasoning"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "query",
          "candidates",
          "citations_per_candidate",
          "critic_verdict",
          "critic_reasoning",
          "trace_id"
        ],
        "title": "QueryResponse",
        "description": "The thing the (eventual) frontend renders."
      },
...
      "SignalBreakdown": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "claim_present": {
            "type": "boolean",
            "title": "Claim Present"
          },
          "self_consistency_score": {
            "type": "number",
            "title": "Self Consistency Score"
          },
          "coherence_score": {
            "type": "number",
            "title": "Coherence Score"
          },
          "peer_anomaly_score": {
            "type": "number",
            "title": "Peer Anomaly Score"
          },
          "trust_score": {
            "type": "number",
            "title": "Trust Score"
          },
          "confidence_interval_low": {
            "type": "number",
            "title": "Confidence Interval Low"
          },
          "confidence_interval_high": {
            "type": "number",
            "title": "Confidence Interval High"
          },
          "confidence_interval_width": {
            "type": "number",
            "title": "Confidence Interval Width"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          },
          "citation_count": {
            "type": "integer",
            "title": "Citation Count"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "claim_present",
          "self_consistency_score",
          "coherence_score",
          "peer_anomaly_score",
          "trust_score",
          "confidence_interval_low",
          "confidence_interval_high",
          "confidence_interval_width",
          "citation_count"
        ],
        "title": "SignalBreakdown"
      },
...
      "TraceTimelineResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TraceTimelineNode"
            },
            "type": "array",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "found",
          "nodes"
        ],
        "title": "TraceTimelineResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}